+++ test/blackbox-tests/mel-new-with-send.t
@@ -18,14 +18,14 @@ Test the interaction between `[@@mel.new]` and `[@@mel.send]`
   
   /*  Not a pure module */
 
-Test `@mel.splice` integration
+Test `@mel.variadic` integration
 
   $ cat > x.ml <<EOF
   > type blue
   > type red
   > external blue : blue = "path/to/blue.js" [@@mel.module]
   > external red : blue -> string array -> red = "Red"
-  >  [@@mel.send] [@@mel.new] [@@mel.splice]
+  >  [@@mel.send] [@@mel.new] [@@mel.variadic]
   > let _ = red blue [| "foo"; "bar" |]
   > EOF
   $ melc -ppx melppx x.ml
@@ -57,14 +57,14 @@ Test `@mel.splice` integration
   
   /*  Not a pure module */
 
-Test `@mel.splice` + `@mel.send.pipe` integration
+Test `@mel.variadic` + `@mel.send.pipe` integration
 
   $ cat > x.ml <<EOF
   > type blue
   > type red
   > external blue : blue = "path/to/blue.js" [@@mel.module]
   > external red : string array -> red = "Red"
-  >   [@@mel.send.pipe: blue] [@@mel.new] [@@mel.splice]
+  >   [@@mel.send.pipe: blue] [@@mel.new] [@@mel.variadic]
   > let _ = blue |> red [| "foo"; "bar" |]
   > EOF
   $ melc -ppx melppx x.ml